fix credit page to actually read effective balance from the durable object
